% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{County names and FIPS codes}
\format{
A data frame with 3,142 rows and 4 variables:
\describe{
  \item{state_abb}{two-letter USPS state abbreviation}
  \item{state_fips}{two-digit state FIPS code}
  \item{county_fips}{three-digit county FIPS code}
  \item{county_name}{name of the county}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/2019/demo/popest/2019-fips.html}
}
\usage{
counties
}
\description{
A data frame containing the names and FIPS codes of counties
in the fifty US states and DC for the year 2019
}
\keyword{datasets}
