% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{State names, abbreviations and FIPS codes}
\format{
A data frame with 51 rows and 3 variables:
\describe{
\item{state_fips}{FIPS code}
\item{state_name}{name of the state}
\item{state_abb}{two-letter USPS state abbreviation}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/2020/demo/popest/2020-fips.html}
}
\usage{
states
}
\description{
A data frame containing the names, USPS abbreviations, and
FIPS codes of the 50 states and DC for the year 2020
}
\keyword{datasets}
