% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urban_area_sf}
\alias{urban_area_sf}
\title{sf data frame of urban areas}
\format{
An sf data frame with 2,613 rows and 4 variables:
\describe{
\item{code}{five-digit urban area code}
\item{name}{urban area name}
\item{states}{vector of states of the urban area}
\item{geometry}{an sf geometry column with polygon data}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.2020.html}
}
\usage{
urban_area_sf
}
\description{
An sf data frame of urban areas in the fifty US states and DC.
It is the 2020 TIGER/Line file.
The EPSG is 4326.
}
\keyword{datasets}
