% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_county_rel}
\alias{zcta_county_rel}
\title{ZCTA-county relationship}
\format{
A data frame with 46,690 rows and 6 variables:
\describe{
\item{zcta}{five-digit ZCTA code}
\item{state_abb}{two-letter USPS state abbreviation}
\item{state_fips}{two-digit state FIPS code}
\item{cty_fips}{three-digit county FIPS code}
\item{subset_area_land}{land area for relationship record, in square meters}
\item{subset_area_land_pct}{percentage of total land area represented by record}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.2020.html#list-tab-E7GGIXHXQT6PO7VYWR}
}
\usage{
zcta_county_rel
}
\description{
A data frame with ZCTA to county relationships as of the 2020 Census.
There is a separate row for each ZCTA-county intersection record.
See \url{https://www2.census.gov/geo/pdfs/maps-data/data/rel2020/zcta520/explanation_tab20_zcta520_county20_natl.pdf}
for additional details.
}
\keyword{datasets}
