% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{County names and FIPS codes}
\format{
A data frame with 3,143 rows and 5 variables:
\describe{
\item{state_abb}{two-letter USPS state abbreviation}
\item{state_fips}{two-digit state FIPS code}
\item{cty_fips}{three-digit county FIPS code}
\item{cty_name}{name of the county}
\item{cty_desc}{county entity type}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/2020/demo/popest/2020-fips.html}
}
\usage{
counties
}
\description{
A data frame containing the names and FIPS codes of counties
in the fifty US states and DC for the year 2020
}
\keyword{datasets}
