% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{place_sf}
\alias{place_sf}
\title{sf data frame of places}
\format{
An sf data frame with 29,321 rows and 4 variables:
\describe{
\item{state_fips}{two-digit state FIPS code}
\item{state_abb}{two-digit USPS abbreviation}
\item{name}{place name}
\item{geometry}{an sf geometry column with polygon data}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}
}
\usage{
place_sf
}
\description{
An sf data frame of places in the fifty US states and DC.
It is the 2019 cartographic boundary file at resolution 500K.
The EPSG is 4326.
}
\keyword{datasets}
