% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_sf}
\alias{zcta_sf}
\title{sf data frame of US Zip Code Tabulation Areas (ZCTA)}
\format{
An sf data frame with 32,989 rows and 5 variables:
\describe{
\item{zcta}{five-digit ZCTA code}
\item{state_abb}{two-digit USPS abbreviation of state containing largest percentage of ZCTA population}
\item{geometry}{an sf geometry column with polygon data}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}
}
\usage{
zcta_sf
}
\description{
An sf data frame of ZCTAs in the fifty US states and DC.
It is the 2019 cartographic boundary file at resolution 500K.
The EPSG is 4326.
}
\keyword{datasets}
